pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vianexus_agent_sdk"
version = "0.1.12"
authors = [
  { name="viaNexus", email="support@vianexus.com" },
]
description = "A Python SDK for the viaNexus AI Agent."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-adk",
    "pydantic<3.0.0",
    "google-generativeai==0.5.4",
    "google.genai==1.25.0",
    "requests==2.32.4",
    "httpx<1.0.0",
    "anyio<5.8.0",
]

[tool.uv.sources]
google-adk = { git = "https://github.com/blueskynexus/adk-python", tag = "v0.1.2-alpha" }

[project.urls]
"Homepage" = "https://github.com/blueskynexus/viaNexus-agent-sdk-python"
"Bug Tracker" = "https://github.com/blueskynexus/viaNexus-agent-sdk-python/issues" 
